feat: split mock and supabase api
